+++ src/RenderPropsComponents/HeaderButton/test.tsx
@@ -1,15 +1,15 @@
 import { shallow } from 'enzyme'
 import React from 'react'
 
-import { BLOCK_CLASS_NAME, NEXT_MONTH_TITLE, PREV_MONTH_TITLE } from '../consts'
-import HeaderButton from '../header_button'
+import { BLOCK_CLASS_NAME, NEXT_MONTH_TITLE, PREV_MONTH_TITLE } from '../../calendar/consts'
+import HeaderButton from '.'
 
 describe('HeaderButton', () => {
   let props, wrapper
 
   beforeEach(() => {
     props = getProps()
-    wrapper = shallow(<HeaderButton {...props} />)
+    wrapper = shallow(<HeaderButton { ...props } />)
   })
 
   it('renders <HeaderButton />', () => {
@@ -50,18 +50,19 @@ describe('HeaderButton', () => {
 
   describe('when prop `arrow` is defined', () => {
     it('renders <HeaderButton />', () => {
-      wrapper.setProps({ arrow: <i className='icon'>Test</i> })
+      wrapper.setProps({
+        arrow: <i className='icon'> Test < /i> })
 
       expect(wrapper).toMatchSnapshot()
+      })
     })
   })
-})
 
-const getProps = (overrides = {}) => ({
-  blockClassName: BLOCK_CLASS_NAME,
-  enabled: true,
-  onClick: () => {},
-  title: PREV_MONTH_TITLE,
-  type: 'prev',
-  ...overrides
-})
+  const getProps = (overrides = {}) => ({
+    blockClassName: BLOCK_CLASS_NAME,
+    enabled: true,
+    onClick: () => { },
+    title: PREV_MONTH_TITLE,
+    type: 'prev',
+    ...overrides
+  })
